fix(mappers): 39 claves ajenas declaraban int contra un bigint
Las 39 apuntan a pcsphp_users.id, que es bigint. No hay migracion: las columnas
reales ya eran bigint(20) tanto en la base como en el .sql que se distribuye. El
que mentia era el mapper, y once modulos tapaban el sintoma con un reemplazo de
cadenas sobre el SQL ya generado.

El unico cambio observable es el DDL: length no interviene, los dos tipos van
juntos en typesInt y en SQLTypesEnum::INTEGERS, nadie compara contra 'int' en
toda la aplicacion, y leida una fila real por el mapper antes y despues el valor
devuelto es identico en las tres columnas que tienen filas.

Las 20 claves ajenas excluidas se comprobaron una a una contra information_schema
—las 20 son int(11)— y los 65 campos int restantes no referencian nada.

scheme-sql-round-trip: 5/8 -> 8/8. PHPStan 859 exactas, sin reparto.

+++ src/app/classes/ApplicationCalls/Mappers/ApplicationCallsMapper.php
@@ -115,7 +115,7 @@ class ApplicationCallsMapper extends EntityMapperExtensible
             'null' => true,
         ],
         'createdBy' => [
-            'type' => 'int',
+            'type' => 'bigint',
             'reference_table' => UsersModel::TABLE,
             'reference_field' => 'id',
             'reference_primary_key' => 'id',
@@ -123,7 +123,7 @@ class ApplicationCallsMapper extends EntityMapperExtensible
             'mapper' => UsersModel::class,
         ],
         'modifiedBy' => [
-            'type' => 'int',
+            'type' => 'bigint',
             'reference_table' => UsersModel::TABLE,
             'reference_field' => 'id',
             'reference_primary_key' => 'id',
